requirements.txt
selenium==4.16.0
pandas==2.1.4
python-dotenv==1.0.0
webdriver-manager==4.0.1
